State Management to Bloc
The State management is moved from Widget layer to Bloc Layer which is search_bloc.dart .
All SetState() is Removed.
A StreamController type bool is Used to change State in bloc .
An abstract Create Method is added to home_page.dart to ensure that HomePage runs with parent Provider of searchBloc.